Split the Flex atom into separate FlexColumn and FlexRow styled atoms.
