+++ src/components/atoms/_base/index.js
@@ -1,5 +1,6 @@
 export { block } from './block'
 export { button, link, navItem, Button, getLinkVariantProps } from './clickable'
 export { control, select } from './control'
+export { flex } from './flex'
 export { grid } from './grid'
 export { list, listTypes, listDefaults, getItemColors } from './list'
