+++ src/components/atoms/atoms.stories.jsx
@@ -10,7 +10,8 @@ export { default as Block } from './block/stories'
 export { default as Button } from './button/stories'
 export { default as Column } from './column/stories'
 export { default as Columns } from './columns/stories'
-export { default as Flex } from './flex/stories'
+export { default as FlexColumn } from './flex-column/stories'
+export { default as FlexRow } from './flex-row/stories'
 export { default as Footer } from './footer/stories'
 export { default as Grid } from './grid/stories'
 export { default as Halves } from './halves/stories'
