+++ src/components/atoms/index.js
@@ -5,7 +5,8 @@ export { default as Block } from './block'
 export { default as Button } from './button'
 export { default as Column } from './column'
 export { default as Columns } from './columns'
-export { FlexColumn, FlexRow } from './flex'
+export { default as FlexColumn } from './flex-column'
+export { default as FlexRow } from './flex-row'
 export { default as Footer } from './footer'
 export { default as Grid } from './grid'
 export { default as Halves } from './halves'
